backend/requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
sentence-transformers
requests
beautifulsoup4
groq
